% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_haplotypeids_hashmap}
\alias{build_haplotypeids_hashmap}
\title{Build hashmap of haplotype ids to individuals}
\usage{
build_haplotypeids_hashmap(individuals, progress = TRUE)
}
\arguments{
\item{individuals}{List of individuals to build hashmap of}

\item{progress}{Show progress?}
}
\value{
Hashmap with haplotype id as keys and vector of individuals as value
}
\description{
Makes it possible to find all individuals with a certain haplotype id.
Must be used with e.g. \code{\link[=get_haplotypeid_matching_individuals_from_hashmap]{get_haplotypeid_matching_individuals_from_hashmap()}}.
}
\seealso{
\code{\link[=get_haplotypeid_matching_individuals_from_hashmap]{get_haplotypeid_matching_individuals_from_hashmap()}}.
}
