% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_haplotypes_hashmap}
\alias{build_haplotypes_hashmap}
\title{Build hashmap of haplotypes to individuals}
\usage{
build_haplotypes_hashmap(individuals, max_load_factor = 100,
  verbose_interval = 0L)
}
\arguments{
\item{individuals}{List of individuals to build hashmap of}

\item{max_load_factor}{Tuning parameter for hash table}

\item{verbose_interval}{0 for no verbose output, e.g. 1,000 for output for every 1000 individual added}
}
\value{
Hashmap with haplotypes as keys and vector of individuals as value
}
\description{
Makes it possible to find all individuals with a certain haplotype.
Must be used with e.g. \code{\link[=get_haplotype_matching_individuals_from_hashmap]{get_haplotype_matching_individuals_from_hashmap()}}
or \code{\link[=print_haplotypes_hashmap]{print_haplotypes_hashmap()}}.
}
\seealso{
\code{\link[=get_haplotype_matching_individuals_from_hashmap]{get_haplotype_matching_individuals_from_hashmap()}}
and \code{\link[=print_haplotypes_hashmap]{print_haplotypes_hashmap()}}.
}
