% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{delete_haplotypeids_hashmap}
\alias{delete_haplotypeids_hashmap}
\title{Delete haplotype id hashmap}
\usage{
delete_haplotypeids_hashmap(hashmap)
}
\arguments{
\item{hashmap}{Hashmap made by \code{\link[=build_haplotypeids_hashmap]{build_haplotypeids_hashmap()}}}
}
\description{
Delete hashmap made by \code{\link[=build_haplotypeids_hashmap]{build_haplotypeids_hashmap()}}.
}
\seealso{
\code{\link[=get_haplotypeid_matching_individuals_from_hashmap]{get_haplotypeid_matching_individuals_from_hashmap()}}
and \code{\link[=build_haplotypeids_hashmap]{build_haplotypeids_hashmap()}}.
}
