% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{delete_haplotypes_hashmap}
\alias{delete_haplotypes_hashmap}
\title{Delete haplotype hashmap}
\usage{
delete_haplotypes_hashmap(hashmap)
}
\arguments{
\item{hashmap}{Hashmap made by \code{\link[=build_haplotypes_hashmap]{build_haplotypes_hashmap()}}}
}
\description{
Delete hashmap made by \code{\link[=build_haplotypes_hashmap]{build_haplotypes_hashmap()}}.
}
\seealso{
\code{\link[=get_haplotype_matching_individuals_from_hashmap]{get_haplotype_matching_individuals_from_hashmap()}}
and \code{\link[=build_haplotypes_hashmap]{build_haplotypes_hashmap()}}.
}
