% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_haplotype_id_individual}
\alias{get_haplotype_id_individual}
\title{Get haplotype id from individual}
\usage{
get_haplotype_id_individual(individual)
}
\arguments{
\item{individual}{Individual to get haplotypes for.}
}
\value{
Haplotype id
}
\description{
Get haplotype id from individual
}
