% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_haplotype_ids_individuals}
\alias{get_haplotype_ids_individuals}
\title{Get haplotype ids from list of individuals}
\usage{
get_haplotype_ids_individuals(individuals)
}
\arguments{
\item{individuals}{Individuals to get haplotypes for.}
}
\value{
List of haplotype ids where element \code{i} is the haplotype of \code{individuals[[i]]}.

[[i]: R:[i
}
\description{
Get haplotype ids from list of individuals
}
