% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_haplotype_matching_individuals_from_hashmap}
\alias{get_haplotype_matching_individuals_from_hashmap}
\title{Get individuals with a certain haplotype by hashmap lookup}
\usage{
get_haplotype_matching_individuals_from_hashmap(hashmap, haplotype)
}
\arguments{
\item{hashmap}{Hashmap to make lookup in, made by \code{\link[=build_haplotypes_hashmap]{build_haplotypes_hashmap()}}}

\item{haplotype}{to get individuals that has this haplotype}
}
\value{
List of individuals with a given haplotype
}
\description{
By using hashmap made by \code{\link[=build_haplotypes_hashmap]{build_haplotypes_hashmap()}},
it is easy to get all individuals with a certain haplotype.
}
\seealso{
\code{\link[=print_haplotypes_hashmap]{print_haplotypes_hashmap()}}
and \code{\link[=build_haplotypes_hashmap]{build_haplotypes_hashmap()}}.
}
