% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_haplotype_no_variants}
\alias{get_haplotype_no_variants}
\title{Get number of variants in haplotype}
\usage{
get_haplotype_no_variants(individual)
}
\description{
Number of variants is for example faster when checking for equality or when
summarising (like plotting). If haplotypes do not have same number of variants, they cannot be equal.
}
