% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_haplotypeid_matching_individuals_from_hashmap}
\alias{get_haplotypeid_matching_individuals_from_hashmap}
\title{Get individuals with a certain haplotype id by hashmap lookup}
\usage{
get_haplotypeid_matching_individuals_from_hashmap(hashmap, haplotype_id)
}
\arguments{
\item{hashmap}{Hashmap to make lookup in, made by \code{\link[=build_haplotypeids_hashmap]{build_haplotypeids_hashmap()}}}

\item{haplotype_id}{to get individuals that has this haplotype id}
}
\value{
List of individuals with a given haplotype id
}
\description{
By using hashmap made by \code{\link[=build_haplotypeids_hashmap]{build_haplotypeids_hashmap()}},
it is easy to get all individuals with a certain haplotype id.
}
\seealso{
\code{\link[=build_haplotypeids_hashmap]{build_haplotypeids_hashmap()}}.
}
