% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_haplotypes_individuals}
\alias{get_haplotypes_individuals}
\title{Get haplotype matrix from list of individuals}
\usage{
get_haplotypes_individuals(individuals)
}
\arguments{
\item{individuals}{Individuals to get haplotypes for.}
}
\value{
Matrix of haplotypes where row `i` is the haplotype of `individuals[[i]]`.
}
\description{
Requires that haplotypes are first populated, e.g. 
with [pedigrees_all_populate_haplotypes()] or 
[pedigrees_all_populate_haplotypes_custom_founders()].
}
\seealso{
[get_haplotypes_pids()].
}
