% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_haplotypes_pids}
\alias{get_haplotypes_pids}
\title{Get haplotypes from a vector of pids.}
\usage{
get_haplotypes_pids(population, pids)
}
\arguments{
\item{population}{Population}

\item{pids}{Vector of pids to get haplotypes for.}
}
\value{
List of haplotypes where row `i` is the haplotype of `individuals[[i]]`.
}
\description{
Requires that haplotypes are first populated, e.g. 
with [pedigrees_all_populate_haplotypes()] or 
[pedigrees_all_populate_haplotypes_custom_founders()].
}
\seealso{
[get_haplotypes_individuals()].
}
