% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_individual_is_female}
\alias{get_individual_is_female}
\title{Is individual female (or male)}
\usage{
get_individual_is_female(individual)
}
\arguments{
\item{individual}{Individual to get sex for.}
}
\value{
Logical: true for female, false for male
}
\description{
Is individual female (or male)
}
