% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_individuals_is_female}
\alias{get_individuals_is_female}
\title{Is individuals females (or males)}
\usage{
get_individuals_is_female(individuals)
}
\arguments{
\item{individuals}{Individuals to get sex for.}
}
\value{
Logical vector: true for female, false for male
}
\description{
Is individuals females (or males)
}
