% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_is_female_in_pedigree}
\alias{get_is_female_in_pedigree}
\title{Get sexes of individuals in pedigree}
\usage{
get_is_female_in_pedigree(ped)
}
\arguments{
\item{ped}{Pedigree to get sexes for}
}
\value{
\code{TRUE} if female, \code{FALSE} if male
}
\description{
Get sexes of individuals in pedigree
}
