% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_matches_info}
\alias{get_matches_info}
\title{Information about matching individuals}
\usage{
get_matches_info(suspect, matching_indv)
}
\arguments{
\item{suspect}{Individual that others must match the profile of.}

\item{matching_indv}{List of matching individuals to get information for.}
}
\value{
Matrix with information about matching individuals.
Columns in order: meioses (meiotic distance to \code{suspect}),
max_L0 (on the path between the matching individual and \code{suspect},
what is the maximum L0 distance between the \code{suspect}'s profile and the
profiles of the individuals on the path),
pid (pid of matching individual)
}
\description{
Note: This function does not check that individuals in
\code{matching_indv} actually match.
}
\details{
This gives detailed information about matching individuals in the pedigree,
e.g. meiotic distances and maximum L0 distance (number of sites they differ) on the path as some of these
matches may have (back)mutations between in between them (but often this will be 0).
}
\seealso{
\code{\link[=count_haplotype_occurrences_individuals]{count_haplotype_occurrences_individuals()}} and \code{\link[=count_haplotype_occurrences_pedigree]{count_haplotype_occurrences_pedigree()}}.
}
