% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{haplotypes_to_hashes}
\alias{haplotypes_to_hashes}
\title{Convert haplotypes to hashes (integers)}
\usage{
haplotypes_to_hashes(haplotypes)
}
\arguments{
\item{haplotypes}{List of haplotypes (list of logical vectors)}
}
\description{
Individuals with the same haplotype will have the same hash (integer)
and individuals with different haplotypes will have different hashes (integers).
}
\details{
This can be useful if for example using haplotypes to define groups
and the haplotype itself is not of interest.
}
