% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{infer_haplotype_ids}
\alias{infer_haplotype_ids}
\title{Infer haplotype ids}
\usage{
infer_haplotype_ids(individuals)
}
\arguments{
\item{individuals}{List of individuals to infer haplotype ids for}
}
\description{
Makes it faster to compare haplotypes by an id instead of the entire mitogenome.
}
