% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitolina.R
\docType{package}
\name{mitolina-package}
\alias{mitolina}
\alias{mitolina-package}
\title{MITOchondrial LINeage Analysis}
\description{
Simulation and tools for analysing mitochondrial lineages.
Genealogies are simulated backwards and binary mtDNA imposed forwards with mutations.
Numerous analyses are possible, e.g. number of matches and meiotic distance to matches.
}
\details{
See vignettes and manual for documentation.

Disclaimer:
THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, TITLE AND NON-INFRINGEMENT.
IN NO EVENT SHALL THE COPYRIGHT HOLDERS OR ANYONE DISTRIBUTING
THE SOFTWARE BE LIABLE FOR ANY DAMAGES OR OTHER LIABILITY, WHETHER IN CONTRACT,
TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/mikldk/mitolina/issues}
}

}
\author{
\strong{Maintainer}: Mikkel Meyer Andersen \email{mikl@math.aau.dk}

}
