% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtdna_mut_Rieux}
\alias{mtdna_mut_Rieux}
\title{Mitogenome mutation rates according to Rieux et al. (2014)}
\format{A tibble / data frame with 5 rows and 6 variables:
\describe{
\item{PartitionRieux}{Partition according to Rieux et al. (2014) (see references)}
\item{MutRateYearly}{Mutation rate per site per year}
\item{MutHPDLower}{Mutation rate highest posterior density lower limit}
\item{MutHPDUpper}{Mutation rate highest posterior density upper limit}
\item{MutRateSEYearly}{Standard error of \code{MutRateYearly} by assuming that the highest posterior region is a 95\% normal}
\item{MutationScheme}{Name of the mutation scheme}
}}
\source{
\url{https://doi.org/10.1093/molbev/msu222}
}
\usage{
mtdna_mut_Rieux
}
\description{
A dataset containing mutation information about the mitogenome
partitioned into 4 partitions
according to Rieux et al. (2014).
}
\details{
Note the value \code{[Not included?]} in the \code{PartitionRieux} column.
}
\references{
\itemize{
\item Rieux et al. (2014): \url{https://doi.org/10.1093/molbev/msu222};
}
}
\seealso{
mtdna_partitions mtdna_mut_Oversti
}
\keyword{datasets}
