% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtdna_mut_RieuxK1}
\alias{mtdna_mut_RieuxK1}
\title{Mitogenome mutation rates according to Rieux et al. (2014)}
\format{A tibble / data frame with 1 row and 5 variables:
\describe{
\item{MutRateYearly}{Mutation rate per site per year}
\item{MutHPDLower}{Mutation rate highest posterior density lower limit}
\item{MutHPDUpper}{Mutation rate highest posterior density upper limit}
\item{MutRateSEYearly}{Standard error of \code{MutRateYearly} by assuming that the highest posterior region is a 95\% normal}
\item{MutationScheme}{Name of the mutation scheme}
}}
\source{
\url{https://doi.org/10.1093/molbev/msu222}
}
\usage{
mtdna_mut_RieuxK1
}
\description{
A dataset containing mutation information about the mitogenome
partitioned into 1 partitions
according to Rieux et al. (2014).
}
\references{
\itemize{
\item PhyloTree: \url{http://www.phylotree.org/resources/rCRS_annotated.htm};
\item Soares et al. (2009): \url{https://dx.doi.org/10.1016/j.ajhg.2009.05.001};
\item Rieux et al. (2014): \url{https://doi.org/10.1093/molbev/msu222};
\item Översti et al. (2017): \url{https://doi.org/10.1038/s41598-017-05673-7}
}
}
\seealso{
mtdna_mut_schemes mtdna_partitions mtdna_mut_RieuxK4 mtdna_mut_SoaresK8 mtdna_mut_OverstiK4
}
\keyword{datasets}
