% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtdna_mut_schemes}
\alias{mtdna_mut_schemes}
\title{Mitogenome with mutation information}
\format{A tibble / data frame with 66,276 (4*16,569) rows and 4 variables:
\describe{
\item{Position}{Position in the mitogenome}
\item{MutationScheme}{Mutation scheme:
"Soares 2009 (K = 8)" \link{mtdna_mut_SoaresK8},
"Rieux 2014 (K = 1)" \link{mtdna_mut_RieuxK1},
"Rieux 2014 (K = 4)" \link{mtdna_mut_RieuxK4}, or
"Översti 2017 (K = 4) 16569" \link{mtdna_mut_OverstiK4}}.
\item{MutRateYearly}{Mutation rate per site per year}
\item{MutRateSEYearly}{Standard error of \code{MutRateYearly}; please refer to
\link{mtdna_mut_SoaresK8}, \link{mtdna_mut_RieuxK1},
\link{mtdna_mut_RieuxK4}, or \link{mtdna_mut_OverstiK4} for details.}
}}
\source{
\url{http://www.phylotree.org/resources/rCRS_annotated.htm}
}
\usage{
mtdna_mut_schemes
}
\description{
A dataset containing information about the mitogenome.
}
\references{
\itemize{
\item PhyloTree: \url{http://www.phylotree.org/resources/rCRS_annotated.htm};
\item Soares et al. (2009): \url{https://dx.doi.org/10.1016/j.ajhg.2009.05.001};
\item Rieux et al. (2014): \url{https://doi.org/10.1093/molbev/msu222};
\item Översti et al. (2017): \url{https://doi.org/10.1038/s41598-017-05673-7}
}
}
\seealso{
mtdna_partitions mtdna_mut_SoaresK8 mtdna_mut_RieuxK4 mtdna_mut_RieuxK1 mtdna_mut_OverstiK4
}
\keyword{datasets}
