% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtdna_mut_SoaresK8}
\alias{mtdna_mut_SoaresK8}
\title{Mitogenome mutation rates according to Soares et al. (2009)}
\format{A tibble / data frame with 8 rows and 4 variables:
\describe{
\item{PartitionSoares}{Partition according to Soares et al. (2009) (see references)}
\item{MutRateYearly}{Mutation rate per site per year}
\item{MutRateSEYearly}{Standard error of \code{MutRateYearly}}
\item{MutationScheme}{Name of the mutation scheme}
}}
\source{
\url{https://dx.doi.org/10.1016/j.ajhg.2009.05.001}
}
\usage{
mtdna_mut_SoaresK8
}
\description{
A dataset containing mutation information about the mitogenome according to
Soares et al. (2009).
Soares et al. (2009) does not give standard errors par partition, but they
write "The substitution rate for the entire molecule was
\eqn{1.665 \times 10^{-8}} (\eqn{\pm 1.479\times 10^{-9}}).".
We assume that the error margin is a 95\% confidence interval for a
normal distribution. This gives a standard error of
\eqn{1.479\times 10^{-9}/1.96 = 7.5459184\times 10^{-10}}.
}
\references{
\itemize{
\item PhyloTree: \url{http://www.phylotree.org/resources/rCRS_annotated.htm};
\item Soares et al. (2009): \url{https://dx.doi.org/10.1016/j.ajhg.2009.05.001};
\item Rieux et al. (2014): \url{https://doi.org/10.1093/molbev/msu222};
\item Översti et al. (2017): \url{https://doi.org/10.1038/s41598-017-05673-7}
}
}
\seealso{
mtdna_mut_schemes mtdna_partitions mtdna_mut_RieuxK4 mtdna_mut_RieuxK1 mtdna_mut_OverstiK4
}
\keyword{datasets}
