% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtdna_vars_Just}
\alias{mtdna_vars_Just}
\title{588 forensic-quality haplotypes representing three U.S. populations}
\format{A tibble / data frame with 588 rows and 5 variables:
\describe{
\item{id}{Sample id}
\item{HG}{Haplogroup}
\item{Variants}{List column with variants for each individual}
\item{Origin}{Origin}
\item{State}{US state}
}}
\source{
\url{https://doi.org/10.1016/j.fsigen.2014.09.021}
}
\usage{
mtdna_vars_Just
}
\description{
Full mtGenome reference data: Development and characterization of 588 forensic-quality
haplotypes representing three U.S. populations
}
\references{
Just et al. "Full mtGenome reference data: Development and characterization of 588
forensic-quality haplotypes representing three U.S. populations", Forensic Science International:
Genetics, 14, 2015.
}
\keyword{datasets}
