% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pedigrees_all_populate_haplotypes}
\alias{pedigrees_all_populate_haplotypes}
\title{Populate haplotypes in pedigrees (founder types same for all).}
\usage{
pedigrees_all_populate_haplotypes(pedigrees, mutation_rates, progress = TRUE)
}
\arguments{
\item{pedigrees}{Pedigree list in which to populate haplotypes}

\item{mutation_rates}{Vector with mutation rates, length \code{loci}}

\item{progress}{Show progress}
}
\description{
Populate haplotypes from founder and down in all pedigrees.
Note, that haplotypes are binary (TRUE/FALSE) and
that all founders get haplotype \code{rep(FALSE, loci)}.
}
\details{
Note, that pedigrees must first have been inferred by \code{\link[=build_pedigrees]{build_pedigrees()}}.
}
\seealso{
\code{\link[=pedigrees_all_populate_haplotypes_custom_founders]{pedigrees_all_populate_haplotypes_custom_founders()}}.
}
