% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{print_haplotypes_hashmap}
\alias{print_haplotypes_hashmap}
\title{Print haplotype hashmap}
\usage{
print_haplotypes_hashmap(hashmap)
}
\arguments{
\item{hashmap}{Hashmap made by \code{\link[=build_haplotypes_hashmap]{build_haplotypes_hashmap()}}}
}
\description{
Print hashmap a haplotypes to individuals made by \code{\link[=build_haplotypes_hashmap]{build_haplotypes_hashmap()}}.
}
\seealso{
\code{\link[=get_haplotype_matching_individuals_from_hashmap]{get_haplotype_matching_individuals_from_hashmap()}}
and \code{\link[=build_haplotypes_hashmap]{build_haplotypes_hashmap()}}.
}
