% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_haplotypes_hashmap}
\alias{build_haplotypes_hashmap}
\title{Build hashmap of haplotypes to individuals}
\usage{
build_haplotypes_hashmap(individuals)
}
\arguments{
\item{individuals}{List of individuals to build hashmap of}
}
\value{
Hashmap with haplotypes as keys and vector of individuals as value
}
\description{
Makes it possible to find all individuals with a certain haplotype.
Must be used with e.g. \code{\link[=get_haplotype_matching_individuals_from_hashmap]{get_haplotype_matching_individuals_from_hashmap()}}
or \code{\link[=print_haplotypes_hashmap]{print_haplotypes_hashmap()}}.
}
\seealso{
\code{\link[=get_haplotype_matching_individuals_from_hashmap]{get_haplotype_matching_individuals_from_hashmap()}}
and \code{\link[=print_haplotypes_hashmap]{print_haplotypes_hashmap()}}.
}
