% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_haplotype_matching_individuals}
\alias{get_haplotype_matching_individuals}
\title{Get individuals matching from list of individuals}
\usage{
get_haplotype_matching_individuals(individuals, haplotype)
}
\arguments{
\item{individuals}{List of individuals to count occurrences in.}

\item{haplotype}{Haplotype to count occurrences of.}
}
\value{
List of individuals that matches \code{haplotype} amongst \code{individuals}.
}
\description{
Get the indvididuals that matches \code{haplotype} in \code{individuals}.
}
