% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtdna_partitions}
\alias{mtdna_partitions}
\title{Mitogenome annotated}
\format{A tibble / data frame with 16,569 rows and 5 variables:
\describe{
\item{Position}{Position in the mitogenome}
\item{rCRS}{The base in the revised Cambridge Reference Sequence (rCRS)}
\item{PartitionPhyloTree}{Partition according to \url{http://www.phylotree.org}}
\item{PartitionRieux}{Partition according to Rieux et al. (2014) (see references)}
\item{PartitionOversti}{Partition according to Rieux et al. (2014) and Översti et al. 2017 (see references)}
}}
\source{
\url{http://www.phylotree.org/resources/rCRS_annotated.htm}
}
\usage{
mtdna_partitions
}
\description{
A dataset containing information about the mitogenome.
}
\details{
Note, that \code{PartitionRieux} and \code{PartitionOversti} have been
modified to obtain the same number of positions as were given in the papers.
}
\references{
\itemize{
\item PhyloTree: \url{http://www.phylotree.org/resources/rCRS_annotated.htm};
\item Rieux et al. (2014): \url{https://doi.org/10.1093/molbev/msu222};
\item Översti et al. (2017): \url{https://doi.org/10.1038/s41598-017-05673-7}
}
}
\seealso{
mtdna_mut_Rieux mtdna_mut_Oversti
}
\keyword{datasets}
